requirements.txt
Keras-Applications==1.0.7
Keras-Preprocessing==1.0.9
Markdown==3.1.1
Werkzeug==0.15.4
absl-py==0.7.1
asn1crypto==0.24.0
astor==0.8.0
certifi==2019.3.9
cffi==1.12.3
cryptography==2.6.1
future==0.17.1
gast==0.2.2
google-pasta==0.1.6
grpcio==1.21.1
h5py==2.9.0
mock==3.0.5
numpy==1.16.3
pbr==5.2.0
pip==19.1.1
protobuf==3.8.0rc1
pycparser==2.19
python-telegram-bot==12.0.0b1
ruamel.yaml==0.15.96
setuptools==41.0.1
six==1.12.0
tb-nightly==1.14.0a20190525
tensorboard==1.13.1
tensorflow==2.0.0a0
tensorflow-estimator==1.13.0
termcolor==1.1.0
tornado==6.0.2
wheel==0.33.4